removed all old type declartions, shared type system fully implemented
